requirements.txt
Flask
gunicorn
pdf2docx
python-pptx
Pillow
PyMuPDF